Complated Custom exception handling including validation exception

+++ LeaveManagement/LeaveManagement.Application/Features/LeaveAllocations/Handler/Commands/DeleteLeaveAllocationHandler.cs
@@ -1,6 +1,7 @@
-﻿using AutoMapper;
-using LeaveManagement.Application.Contracts.Persistence;
+﻿using LeaveManagement.Application.Contracts.Persistence;
+using LeaveManagement.Application.Exceptions;
 using LeaveManagement.Application.Features.LeaveAllocations.Request.Commands;
+using LeaveManagement.Domain;
 using MediatR;
 
 namespace LeaveManagement.Application.Features.LeaveAllocations.Handler.Commands
@@ -8,17 +9,20 @@ namespace LeaveManagement.Application.Features.LeaveAllocations.Handler.Commands
     public class DeleteLeaveAllocationHandler : IRequestHandler<DeleteLeaveAllocation>
     {
         private readonly ILeaveAllocationRepository _leaveAllocationRepository;
-        private readonly IMapper _mapper;
 
-        public DeleteLeaveAllocationHandler(ILeaveAllocationRepository leaveAllocationRepository, IMapper mapper)
+        public DeleteLeaveAllocationHandler(ILeaveAllocationRepository leaveAllocationRepository)
         {
             _leaveAllocationRepository = leaveAllocationRepository;
-            _mapper = mapper;
         }
         public async Task<Unit> Handle(DeleteLeaveAllocation request, CancellationToken cancellationToken)
         {
             var leaveAllocation = await _leaveAllocationRepository.Get(request.Id);
+
+            if (leaveAllocation == null)
+                throw new NotFoundException(nameof(LeaveAllocation), request.Id);
+
             await _leaveAllocationRepository.Delete(leaveAllocation);
+
             return Unit.Value;
         }
     }
