+++ LeaveManagement/LeaveManagement.Application/Features/LeaveTypes/Handler/Commands/DeleteleaveTypeHandler.cs
@@ -1,5 +1,5 @@
-﻿using AutoMapper;
-using LeaveManagement.Application.Contracts.Persistence;
+﻿using LeaveManagement.Application.Contracts.Persistence;
+using LeaveManagement.Application.Exceptions;
 using LeaveManagement.Application.Features.LeaveTypes.Request.Commands;
 using MediatR;
 
@@ -8,16 +8,18 @@ namespace LeaveManagement.Application.Features.LeaveTypes.Handler.Commands
     public class DeleteleaveTypeHandler : IRequestHandler<DeleteleaveType>
     {
         private readonly ILeaveTypeRepository _leaveTypeRepository;
-        private readonly IMapper _mapper;
 
-        public DeleteleaveTypeHandler(ILeaveTypeRepository leaveTypeRepository, IMapper mapper)
+        public DeleteleaveTypeHandler(ILeaveTypeRepository leaveTypeRepository)
         {
             _leaveTypeRepository = leaveTypeRepository;
-            _mapper = mapper;
         }
         public async Task<Unit> Handle(DeleteleaveType request, CancellationToken cancellationToken)
         {
             var leaveType = await _leaveTypeRepository.Get(request.Id);
+
+            if (leaveType == null)
+                throw new NotFoundException(nameof(LeaveTypes), request.Id);
+
             await _leaveTypeRepository.Delete(leaveType);
 
             return Unit.Value;
